li voleva interi

+++ esame.py
@@ -33,7 +33,7 @@ class Diff():
 
         result = []
         for i in range(len(data) - 3):
-                res = (data[i+1] - data[i])/self.ratio
+                res = int((data[i+1] - data[i])/self.ratio)
                 result.append(res) 
                 
         return result
